\name{hann-package}
\alias{hann-package}
\alias{hann}
\docType{package}
\title{
Hopfield Artificial Neural Networks
}
\description{
  \pkg{hann} provides tools to build Hopfield-based artificial neural
  networks. Two types of networks can be built: one-layer Hopfield
  network, and three-layer network with a hidden (convoluted) layer.

  The complete list of functions can be displayed with
  \code{library(help = hann)}. See the vignette
  ``IntroductionHopfieldNetworks'' for several examples.

  More information on \pkg{hann} can be found at
  \url{https://github.com/emmanuelparadis/hann}.
}

\author{
  Emmanuel Paradis

  Maintainer: Emmanuel Paradis <Emmanuel.Paradis@ird.fr>
}
\keyword{package}
